perf: switch Perlin noise to f32 and hoist row-invariant computation
- All noise arithmetic now uses f32 (~2x throughput on ARM NEON)
- Row-level y-axis values (grid cell, fractional offset, fade curve) are
  computed once per row and reused across all 1830 columns
- Permutation lookups for p[xi] and p[xi+1] hoisted out of hash step
- f32 precision is more than sufficient for UInt8 pixel output

+++ spatialbench-raster/src/cog/mod.rs
@@ -46,7 +46,7 @@ pub struct CogConfig {
     /// Internal tile size (pixels per side).
     pub tile_size: u32,
     /// Perlin noise frequency (controls spatial detail per tile).
-    pub noise_frequency: f64,
+    pub noise_frequency: f32,
 }
 
 impl Default for CogConfig {
